+++ extension/frontend/view/old/flare.js
