--- extension/frontend/view/reactview.js
@@ -1,186 +1,0 @@
-import React, { useEffect, useState,useRef } from 'react';
-import ReactDOM from 'react-dom';
-import {
-    BrowserRouter,
-    Switch,
-    Route,
-    Link,
-    Router
-  } from "react-router-dom";
-import { createTypeReferenceNode } from 'typescript';
-import * as d3 from '../../libraries/d3.js';
-
-
-
-const App = () => {
-    return (
-        <div>
-            <Container />
-        </div>
-    )
-}
-
-const port = chrome.runtime.connect({name: 'test'})
-
-const Container = () => {
-
-    const [tree, setTree] = useState()
-
-    // createTree((tree) => {
-
-    // })
-
-    useEffect(() => {
-        port.postMessage({
-        name:'connect',
-        tabID: chrome.devtools.inspectedWindow.tabId,
-        })
-
-        port.onMessage.addListener((message) => {
-            if (message.length === 2){
-                // console.log('message received by panel', message);
-                setTree(message)
-            }
-        })
-    },[]) 
-
-   
-    return (
-        <div>
-            <Tree tree={tree} />
-            <Atoms tree={tree}/>
-        </div>
-    )
-}
-
-
-
-
-const Tree = (props) => {
-    useEffect(() => {
-        if (props.tree){
-            console.log('props tree',props.tree[0])
-
-            document.querySelector('#canvas').innerHTML = ''
-
-            const root = d3.hierarchy(props.tree[0])
-
-            const panelWidth = Math.floor(window.innerWidth * 0.66);
-
-
-            // Find out the height of the tree and size the svg accordingly (each level havin 95px)
-            const dataHeight = root.height;
-            const treeHeight = dataHeight * 95;
-            // console.log('windowHeight:',window.innerHeight,'treeHeight:', treeHeight,'dataHeight',dataHeight)
-            const svgHeight = Math.max(window.innerHeight, treeHeight)
-
-            const svg = d3.select('#canvas')
-                .append('svg')
-                .attr('width', panelWidth)
-                .attr('height', svgHeight)
-                .append('g')
-                .attr('transform', 'translate(-30, 30)');
-
-            let tree = d3.tree().size([panelWidth-80, treeHeight]);
-            tree(root)
-
-            const nodes = root.descendants()
-           
-            const node = svg.selectAll('.node')
-                .data(nodes)
-                .enter()
-                .append('g')
-                .attr('class','node')
-                .attr('transform', (d) => 'translate(' + d.x + ',' + d.y + ')')
-
-                node.append('circle')
-                .attr('r',5)
-                .attr('fill','steelblue')
-
-                node.append('text')
-                .attr('x', 10)
-                    .attr('y', 5)
-                .text(function(d) {
-                return d.data.name
-                })
-
-            const links = root.links()
-            const link = svg.selectAll('.link')
-                  .data(links)
-                  .join('path')
-                  .attr('class','link')
-                  .attr('d', d3.linkVertical()
-                    .x(d => d.x)
-                    .y(d => d.y))
-                    .attr("fill", "none") 
-                    .attr("stroke", "green")
-                    .attr("stroke-opacity", 0.4)
-                    .attr("stroke-width", 1.5)
-        }
-    })
-
-    return (
-        <div>
-            <div id='canvas'></div>
-        </div>
-    )
-}
- 
-
-const Atoms = (props) => {
-    useEffect(() => {
-        if (props.tree){
-            const masterContainer = document.querySelector('#atoms')
-            masterContainer.innerHTML = ''
-
-            const atoms = props.tree[1].atomVal
-            for (let prop in atoms){
-                let atomContainer = document.createElement('div')
-                atomContainer.classList.add('atom-div')
-                
-                console.log(prop)
-                // Set atomname text to prop which will be the name of the atom
-                const atomName = document.createElement('h3')
-                atomName.textContent = prop
-              
-                let originalColor;
-
-                atomName.addEventListener('click',(e) => {
-                    // console.log('is it working d3 blink')
-                    let atom = e.target.innerHTML
-                      d3.selectAll('circle').style('fill', originalColor)
-                    
-                    let svg = d3.select('#canvas').selectAll('.node').each(function(e){
-                      if (!e.data.atoms){}
-                      else if(e.data.atoms.includes(atom)){
-                        d3.select(this).select('circle').style('fill','#eee')
-                      }
-                      
-                    })
-                  })
-
-                let atomState = document.createElement('h3')
-                atomState.classList.add('s-margin')
-
-
-                let text = JSON.stringify(atoms[prop])
-                atomState.textContent = `State : ${text}`
-
-
-                atomContainer.appendChild(atomName)
-                atomContainer.appendChild(atomState)
-                masterContainer.appendChild(atomContainer)
-            }
-          
-        }
-    })
-
-    return (
-        <div>
-            <div id='atoms'></div>
-            <h1>ATOMS TEST</h1>
-        </div>
-    )
-}
-ReactDOM.render(<App />, document.getElementById('root'))
-
