+++ extension/frontend/view/view.html
@@ -8,22 +8,6 @@
 </head>
 <body>
    <div id="root"></div>
-    <!-- <div id="total-recoil">Total RecoilJS</div>
-    <div id="currentComponent" >No Atoms Hovered</div>
-    <div id="main-container">
-      <div id="tree"></div>
-      <div id="info">
-        <div class="options">
-          <button>Current</button>
-          <button>History</button>
-        </div>
-        <div class='information'></div>
-         <h1>Node Information</h1>
-        <h1 id="state-info"></h1> -->
-      <!-- </div>
-    </div>
-  
-  </div> -->
     <script src="bundle.js"></script>
 </body>
 </html>
